Updated proto to 0.5.0
The update to proto introduced some breaking changes (see WikiOpenCite/proto#24) along with some datatype changes in WikiOpenCite/proto#25 that meant some changes to the interface citescoop provides.

Instead of returning a revision map now, the dump parser will instead return a map (basically the same thing just no longer wrapped in a protobuf object).

+++ src/parser/parser_impl.cc
@@ -104,16 +104,16 @@ proto::ExtractedCitation Parser::ParserImpl::BuildCitation(
             algo::trim_copy(param.value.value()));
       } else if (key == "pmid") {
         try {
-          citation.mutable_identifiers()->set_pmid(
-              this->StrToIntIdent(algo::trim_copy(param.value.value())));
+          citation.mutable_identifiers()->set_pmid(static_cast<uint32_t>(
+              this->StrToIntIdent(algo::trim_copy(param.value.value()))));
         } catch (const TemplateParseException& e) {
           if (!this->options().ignore_invalid_ident)
             throw e;
         }
       } else if (key == "pmc") {
         try {
-          citation.mutable_identifiers()->set_pmcid(
-              this->ParsePmcId(algo::trim_copy(param.value.value())));
+          citation.mutable_identifiers()->set_pmcid(static_cast<uint32_t>(
+              this->ParsePmcId(algo::trim_copy(param.value.value()))));
         } catch (const TemplateParseException& e) {
           if (!this->options().ignore_invalid_ident)
             throw e;
